refactor: split out artifacts from main
