+++ dapp/src/lib.rs
@@ -1,6 +1,6 @@
 use ethers::{
-    abi::{self},
-    types::*,
+    abi,
+    types::{Address, U256},
     utils::{keccak256, CompiledContract},
 };
 
@@ -19,6 +19,9 @@ use executor::{Executor, MemoryState};
 
 mod runner;
 use runner::{ContractRunner, TestResult};
+
+mod artifacts;
+use artifacts::DapptoolsArtifact;
 
 /// Re-export of the Rust EVM for convenience
 pub use evm;
@@ -56,65 +59,6 @@ pub struct MultiContractRunner<'a> {
     pub init_state: MemoryState,
     pub state: MemoryState,
     pub gas_limit: u64,
-}
-
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct DapptoolsArtifact {
-    contracts: HashMap<String, HashMap<String, serde_json::Value>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct Contract {
-    abi: ethers::abi::Abi,
-    evm: Evm,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct Evm {
-    bytecode: Bytecode,
-    deployed_bytecode: Bytecode,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct Bytecode {
-    #[serde(deserialize_with = "deserialize_bytes")]
-    object: Bytes,
-}
-
-use serde::Deserializer;
-
-pub fn deserialize_bytes<'de, D>(d: D) -> Result<Bytes, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value = String::deserialize(d)?;
-
-    Ok(hex::decode(&value)
-        .map_err(|e| serde::de::Error::custom(e.to_string()))?
-        .into())
-}
-
-impl DapptoolsArtifact {
-    fn contracts(&self) -> Result<HashMap<String, CompiledContract>> {
-        let mut map = HashMap::new();
-        for (key, value) in &self.contracts {
-            for (contract, data) in value.iter() {
-                let data: Contract = serde_json::from_value(data.clone())?;
-                let data = CompiledContract {
-                    abi: data.abi,
-                    bytecode: data.evm.bytecode.object,
-                    runtime_bytecode: data.evm.deployed_bytecode.object,
-                };
-                map.insert(format!("{}:{}", key, contract), data);
-            }
-        }
-
-        Ok(map)
-    }
 }
 
 impl<'a> MultiContractRunner<'a> {
@@ -245,7 +189,7 @@ impl<'a> MultiContractRunner<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use ethers::{prelude::Lazy, utils::id};
+    use ethers::{prelude::Lazy, utils::id, types::H160};
 
     use dapp_utils::get_func;
     use evm::{ExitReason, ExitRevert, ExitSucceed};
